move sub_stat_edict
