+++ nni/nas/strategy/evolution.py
@@ -200,128 +200,3 @@ class RandomEvolution(RegularizedEvolution):
                 break
 
 
-dict2tuple = lambda d: tuple(d[k] for k in sorted(d.keys()))
-
-class AutoformerEvolution(BaseStrategy):
-    """
-    Algorithm for Autoformer evolution.
-    Follows evolution search in Chen et al. "AutoFormer: Searching Transformers for Visual Recognition".
-    Parameters
-    ----------
-    mutation_size : int
-        The number of individuals that should mutate in each generation. Default: 25.
-    crossover_size: int
-        The number of individuals that should cross over in each generation. Default: 25.
-    mutation_prob : float
-        Probability that mutation happens default. Default: 0.4
-    depth_mutation_prob : float
-        Probability that mutation happens in depth. Default: 0.2
-    """
-    MAX_ITER = 100
-    def __init__(self, generations=20, population_size=50, parent_size=25, 
-            mutation_size=25, crossover_size=25, mutation_prob=0.4, depth_mutation_prob=0.2, 
-            optimize_mode='maximize', on_failure='ignore', model_filter=None):
-        assert parent_size < population_size and crossover_size + mutation_size <= population_size
-        assert generations > 1
-        super().__init__(optimize_mode=optimize_mode, population_size=population_size, sample_size=25, cycles=100,
-                 mutation_prob=0.1, on_failure=on_failure, model_filter=model_filter)
-
-        self.generations = generations
-        self.parent_size = parent_size
-        self.mutation_size = mutation_size
-        self.crossover_size = crossover_size
-        self.depth_mutation_prob = depth_mutation_prob
-        self.mutation_prob = mutation_prob
-
-        self._cache = set()
-
-    def _cache_hit(self, config):
-        if dict2tuple(config) in self._cache:
-            return True
-        else:
-            self._cache.add(config)
-            return False
-
-    def random(self, search_space):
-        for _ in range(self.MAX_ITER):
-            config = {k: random.choice(v) for k, v in search_space.items()}
-            if not self._cache_hit(config):
-                break
-        return config
-
-    def mutation(self, parents, search_space):
-        for _ in range(self.MAX_ITER):
-            config = {k: v for k, v in random.choice(parents).item()}    # copy
-            # mutate depth
-            if random.random() < self.depth_mutation_prob:
-                config["depth"] = random.choice(search_space["depth"])
-            # mutate embed_dim
-            if random.random() < self.mutation_prob:
-                config["embed_dim"] = random.choice(search_space["embed_dim"])
-            # mlp_ratio
-            for d in range(config["depth"]):
-                if random.random() < self.mutation_prob:
-                    config[f'mlp_ratio_{d}'] = random.choice(search_space[f'mlp_ratio_{d}'])
-                if random.random() < self.mutation_prob:
-                    config[f'num_head_{d}'] = random.choice(search_space[f'num_head_{d}'])
-
-            if not self._cache_hit(config):
-                break
-
-        return config
-
-    def crossover(self, parents, search_space):
-        p1 = random.choice(parents)
-        for _ in range(self.MAX_ITER):
-            p2 = random.choice(parents)
-            if p1["depth"] != p2["depth"]:
-                continue
-            config = {k: v for k, v in p1.item()}    # copy
-            for k in p1.keys():
-                config[k] = random.choice([p1[k], p2[k]])
-            if not self._cache_hit(config):
-                break
-
-        return config
-
-    def topk_parents(self, parents):
-        reverse = self.optimize_mode=='maximize'
-        parents = sorted(parents, key = lambda p: p.y, reverse = reverse)
-        return [p.x for p in parents]
-
-    def run(self, base_model, applied_mutators):
-        search_space = dry_run_for_search_space(base_model, applied_mutators)
-        # Run the first population regardless concurrency
-        _logger.info('Initializing the first generations.')
-        while True:
-            # try to submit new models
-            while len(self._population) + len(self._running_models) < self.population_size:
-                config = self.random(search_space)
-                self._submit_config(config, base_model, applied_mutators)
-            # collect results
-            self._move_succeeded_models_to_population()
-            self._remove_failed_models_from_running_list()
-            time.sleep(self._polling_interval)
-
-            if len(self._population) >= self.population_size:
-                break
-
-        for _ in range(self.generations - 1):
-            parents = self.topk_parents([self._population])
-            population = self.mutate(parents, search_space) + self.crossover(parents, search_space)
-            while len(population) < self.population_size:
-                config = self.random(search_space)
-                population.append(config)
-
-            while True:
-                # try to submit new models
-                while query_available_resources() > 0 and len(population) > 0:
-                    config = population.pop()
-                    self._submit_config(config, base_model, applied_mutators)
-                # collect results
-                self._move_succeeded_models_to_population()
-                self._remove_failed_models_from_running_list()
-                time.sleep(self._polling_interval)
-
-                if len(self._running_models) + len(population) == 0:
-                    break
